Fix issue where trivial log messages are output wrapped in quotes
Serilog, by default, wraps string properties in quotes to communicate the underlying data type. Adding the Serilog specific 'l' format string to the property instructs Serilog to treat the property as a literal string and output the value without wrapping it in quotes.

Since we are wrapping trivial log messages in a single token message template (as is best practice in this case to avoid thrashing the message template cache), it makes sense to drop the surrounding quotes. The only purpose of the wrapping quotes is to communicate type information in the context of meaningful log event property, and we are wrapping a message in a token with the understand that our input has no meaniful event properties.